requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
psycopg2-binary==2.9.10
python-dotenv==1.2.2
Werkzeug==3.1.8
requests==2.33.1
